moving review handlers into own folder and separating out handler
